requirements.txt
numpy==1.14.3
opencv-python==3.4.4.19
scipy==1.1.0
tqdm==4.28.1
Pillow==5.3.0
scikit-learn==0.18.1
tensorboardX==1.5